+++ service/controller/resources/vaultrole/current.go
@@ -6,7 +6,7 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/vaultrole"
 
-	"github.com/giantswarm/cert-operator/service/controller/v2/key"
+	"github.com/giantswarm/cert-operator/service/controller/key"
 )
 
 func (r *Resource) GetCurrentState(ctx context.Context, obj interface{}) (interface{}, error) {
